This PR introduces a new loaded keyspace event
The new event will be fired after a key was loaded from RDB and added to the keyspace.
The new event will be triggered to modules which registered to REDISMODULE_NOTIFY_GENERIC event type.

This PR also introduce tests for the new notification inside a new test file called: keyspace_events.c